added strategies for all conflict resolution types
