--- saved-sites/saved-sites-impl/src/main/java/com/duckduckgo/savedsites/impl/sync/algorithm/SavedSitesTimestampAlgorithm.kt
@@ -1,45 +1,0 @@
-/*
- * Copyright (c) 2023 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.savedsites.impl.sync.algorithm
-
-import com.duckduckgo.savedsites.impl.sync.SyncBookmarkEntry
-
-class SavedSitesTimestampAlgorithm() : SavedSitesSyncAlgorithm {
-    override fun processBookmarkFolder(
-        remoteFolder: SyncBookmarkEntry,
-        parentId: String,
-        lastModified: String
-    ) {
-        TODO("Not yet implemented")
-    }
-
-    override fun processFavouritesFolder(
-        entries: List<SyncBookmarkEntry>,
-        lastModified: String
-    ) {
-        TODO("Not yet implemented")
-    }
-
-    override fun processBookmark(
-        child: String,
-        entries: List<SyncBookmarkEntry>,
-        folderId: String,
-        lastModified: String
-    ) {
-        TODO("Not yet implemented")
-    }
-}
